+++ packages/react/src/Active.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { ReactElement } from 'react';
 import PropTypes from 'prop-types';
 import invariant from 'invariant';
-import { CuriConfig, Response } from '@curi/core';
+import { CuriRouter, Response } from '@curi/core';
 import { HickoryLocation } from '@hickory/root';
 
 export interface ActiveProps {
@@ -13,14 +13,14 @@ export interface ActiveProps {
   merge(props: object): object;
   extra?(l: HickoryLocation, d: object): boolean;
   details?: object;
-  curi?: CuriConfig;
+  curi?: CuriRouter;
   response?: Response;
 }
 
 class Active extends React.Component<ActiveProps, {}> {
   static contextTypes = {
     curi: PropTypes.shape({
-      config: PropTypes.object,
+      router: PropTypes.object,
       response: PropTypes.object
     })
   };
@@ -30,16 +30,16 @@ class Active extends React.Component<ActiveProps, {}> {
   }
 
   verifyActiveAddon() {
-    const curi = this.props.curi || this.context.curi.config;
+    const curi = this.props.curi || this.context.curi.router;
     invariant(
       curi.addons.active,
       'You are attempting to use the "active" prop, but have not included the "active" ' +
-        'addon (curi-addon-active) in your Curi configuration object.'
+        'addon (curi-addon-active) in your Curi router.'
     );
   }
 
   render() {
-    const curi = this.props.curi || this.context.curi.config;
+    const curi = this.props.curi || this.context.curi.router;
     const response = this.props.response || this.context.curi.response;
     const {
       extra,