+++ packages/react/src/CuriBase.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { CuriConfig, Response } from '@curi/core';
+import { CuriRouter, Response } from '@curi/core';
 import { CuriContext } from './interface';
 import { Action } from '@hickory/root';
 
 export interface CuriBaseProps {
-  config: CuriConfig;
+  router: CuriRouter;
   render: (
     r: Response,
     action: string,
-    c?: CuriConfig
+    c?: CuriRouter
   ) => React.ReactElement<any>;
   response: Response;
   action?: Action;
@@ -18,7 +18,7 @@ export interface CuriBaseProps {
 class CuriBase extends React.Component<CuriBaseProps> {
   static childContextTypes = {
     curi: PropTypes.shape({
-      config: PropTypes.object,
+      router: PropTypes.object,
       response: PropTypes.object,
       action: PropTypes.string
     })
@@ -31,7 +31,7 @@ class CuriBase extends React.Component<CuriBaseProps> {
   getChildContext(): CuriContext {
     return {
       curi: {
-        config: this.props.config,
+        router: this.props.router,
         response: this.props.response,
         action: this.props.action
       }
@@ -42,7 +42,7 @@ class CuriBase extends React.Component<CuriBaseProps> {
     return this.props.render(
       this.props.response,
       this.props.action,
-      this.props.config
+      this.props.router
     );
   }
 }