+++ packages/react/src/interface.ts
@@ -1,4 +1,4 @@
-import { CuriConfig, Response } from '@curi/core';
+import { CuriRouter, Response } from '@curi/core';
 import { Action } from '@hickory/root';
 
 export interface CuriContext {
@@ -6,7 +6,7 @@ export interface CuriContext {
 }
 
 export interface CuriProps {
-  config: CuriConfig;
+  router: CuriRouter;
   response: Response;
   action: Action;
 }