+++ packages/react/types/Block.d.ts
@@ -2,11 +2,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { ConfirmationFunction } from '@hickory/root';
-import { CuriConfig } from '@curi/core';
+import { CuriRouter } from '@curi/core';
 export interface BlockProps {
     active?: boolean;
     confirm: ConfirmationFunction;
-    curi?: CuriConfig;
+    curi?: CuriRouter;
 }
 declare class Block extends React.Component<BlockProps> {
     static contextTypes: {