+++ packages/react/types/CuriBase.d.ts
@@ -1,12 +1,12 @@
 /// <reference types="react" />
 import React from 'react';
 import PropTypes from 'prop-types';
-import { CuriConfig, Response } from '@curi/core';
+import { CuriRouter, Response } from '@curi/core';
 import { CuriContext } from './interface';
 import { Action } from '@hickory/root';
 export interface CuriBaseProps {
-    config: CuriConfig;
-    render: (r: Response, action: string, c?: CuriConfig) => React.ReactElement<any>;
+    router: CuriRouter;
+    render: (r: Response, action: string, c?: CuriRouter) => React.ReactElement<any>;
     response: Response;
     action?: Action;
 }