+++ packages/react/types/interface.d.ts
@@ -1,10 +1,10 @@
-import { CuriConfig, Response } from '@curi/core';
+import { CuriRouter, Response } from '@curi/core';
 import { Action } from '@hickory/root';
 export interface CuriContext {
     curi: CuriProps;
 }
 export interface CuriProps {
-    config: CuriConfig;
+    router: CuriRouter;
     response: Response;
     action: Action;
 }